Split specific evaluation in new module
